+++ src/main/scala/com/johnsnowlabs/ml/ai/util/Generation/Logit/LogitWarper/TopKLogitWarper.scala
@@ -23,7 +23,7 @@ class TopKLogitWarper(
     val minTokensToKeep: Int = 1)
     extends LogitWarper {
   override def call(
-      inputIds: Seq[Array[Long]],
+      inputIds: Seq[Array[Int]],
       scores: Array[Array[Float]],
       currentLength: Int): Array[Array[Float]] = {
     var logitsUpd = scores
